% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_drafts.R
\name{get_drafts}
\alias{get_drafts}
\title{Gets draft information, players' names, and player URLs for specified draft and season}
\usage{
get_drafts(draft_type, draft_year, progress = TRUE, other = "", ...)
}
\arguments{
\item{draft_type}{The type of draft for which the user wants to scrape data. Draft types must be typed exactly as they are found on EliteProspects (though case doesn't matter). Draft types include -- but are not limited to -- NHL Entry Draft, NHL Expansion Draft, KHL Draft, NWHL Draft, CWHL Draft, CHL Import Draft, and OHL U18 Priority Selection. Others may be found at the bottom of the page at \url{https://www.eliteprospects.com/draft/nhl-entry-draft}.}

\item{draft_year}{Seasons for which the user wants to scrape data. Must be of the form \code{2018}, \code{1996}, etc -- only  a single 4-digit number.}

\item{progress}{Sets a Progress Bar. Defaults to \code{TRUE}.}

\item{...}{Allows the user to supply other information to the function. If you don't know what this means, then don't worry about it.}
}
\description{
Returns a data frame of draft information, players' names, and player URLs for user supplied drafts (NHL Entry Draft, CHL Import Draft, etc.) & seasons.
}
\examples{
get_drafts("chl import draft", 2018)

get_drafts(c("nhl entry draft", "khl draft"), c(1994, 2017:2018))

}
