% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_stats_team.R
\name{get_player_stats_team}
\alias{get_player_stats_team}
\title{Returns a data frame of players' stats and player URLs for user supplied teams & seasons}
\usage{
get_player_stats_team(..., .progress = FALSE)
}
\arguments{
\item{...}{Function requires a "team_url", "team_name", "league", and "season." Additional data may be supplied. All of this information comes directly from "get_teams()", if desired.}

\item{.progress}{Sets a Progress Bar. Defaults to FALSE.}
}
\description{
Returns a data frame of players' stats and player URLs for user supplied teams & seasons
}
\examples{

# The function works in conjunction with get_teams()
teams <- get_teams("ohl", "2017-2018")
get_player_stats_team(teams)

# All functions are easily pipeable too
get_teams(c("shl", "allsvenskan"), c("2008-2009", "2009-2010", "2010-2011")) \%>\%
  get_player_stats_team(.progress = TRUE)
  
# It's also easy to get player stats for only 1 team   
get_teams("ncaa iii", "2017-2018") \%>\%
  filter(team == "Hamilton College") \%>\%
  get_player_stats_team()
  
}
