% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{Gets teams and team URLs for specified league and season}
\usage{
get_teams(league, season, progress = TRUE, other = "", ...)
}
\arguments{
\item{league}{Leagues from which the user wants to scrape data}

\item{season}{Seasons for which the user wants to scrape data. Must be of the form \code{2018}, \code{1965}, etc. for the 2017-18 and 1964-65 seasons, respectively.}

\item{progress}{Sets a Progress Bar. Defaults to \code{TRUE}.}

\item{...}{Allows the user to supply other information to the function. If you don't know what this means, then don't worry about it.}
}
\description{
Returns a data frame of teams and their URLs for user supplied leagues & seasons. Bear in mind that there are some cases in which teams that aren't part of the user-supplied league are returned in the data frame. This is normal and is fixed later when using \code{get_player_stats_team()}.
}
\examples{
get_teams("ohl", 2013)

get_teams(c("SHL", "allsvenskan", "ncaa iii"), c(1994, 2017:2018), progress = FALSE)

}
