% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_years_difference.R
\name{get_years_difference}
\alias{get_years_difference}
\title{Gets exact age}
\usage{
get_years_difference(from, to)
}
\arguments{
\item{from}{The start date. Typically the player's birth date.}

\item{to}{The reference date. Used for finding a player's draft eligibility, it's typically \code{2018-09-15}, or whatever the year of the NHL draft is.}
}
\description{
Returns the exact age of a player given his birth date and some reference date
}
\examples{
get_years_difference("2000-09-15", "2018-09-15")

# My birthday -- don't forget it!
get_years_difference("1995-10-22", "2018-08-27")

}
